+++ src/index.ts
@@ -6,6 +6,8 @@ import * as Environments from "./Environment";
 import { isRegExp } from "util/types";
 import { inspect } from "util";
 import * as os from "os";
+import * as fs from "fs";
+import path from "path";
 
 class Utility extends Environments.ProtectedEnvironment {
     public constructor() {
@@ -77,8 +79,83 @@ class Utility extends Environments.ProtectedEnvironment {
         this.set("throw", (handler) => {
             throw handler.waitForArguments(handler.getArg(0)).shift();
         });
+        this.set("eval", (handler) => {
+            if (handler.argLength < 1) throw new Error("Expected 3 arguments, instead got " + handler.argLength + "!");
+            let [evalStr, env=handler.getEnv(), returnOut=false] = handler.waitForArguments(...handler.getArgs(0, 3)) ?? [];
+            if (typeof evalStr !== "string") throw new Error("arg0 must be typeof string!");
+            let script = new Scripts.Script("eval", evalStr, env);
+            script.prepareModules();
+            let out = script.run();
+            if (returnOut) return out;
+        });
 
         this.set("Error", () => Error);
+
+        this.cache.lock();
+    }
+}
+
+class Path extends Environments.ProtectedEnvironment {
+    public constructor() {
+        super();
+
+        this.set("resolve", (handler) => {
+            return path.resolve(...handler.waitForArguments(...handler.getArgs(0, handler.argLength)));
+        });
+        this.set("join", (handler) => {
+            return path.join(...handler.waitForArguments(...handler.getArgs(0, handler.argLength)));
+        });
+        this.set("isAbsolute", (handler) => {
+            return path.isAbsolute(handler.waitForArguments(...handler.getArgs(0, 1)).shift());
+        });
+        this.set("dirname", (handler) => {
+            return path.dirname(handler.waitForArguments(...handler.getArgs(0, 1)).shift());
+        });
+        this.set("basename", (handler) => {
+            return path.basename(handler.waitForArguments(...handler.getArgs(0, 1)).shift());
+        });
+        this.set("extname", (handler) => {
+            return path.extname(handler.waitForArguments(...handler.getArgs(0, 1)).shift());
+        });
+
+        this.cache.lock();
+    }
+}
+
+class FileSystem extends Environments.ProtectedEnvironment {
+    public constructor() {
+        super();
+
+        this.set("readdirSync", (handler) => {
+            return fs.readdirSync(handler.waitForArguments(handler.getArg(0)).shift());
+        });
+        this.set("readFileSync", (handler) => {
+            let args = handler.waitForArguments(...handler.getArgs(0,2))
+            return fs.readFileSync(args.shift(), args.shift());
+        });
+        this.set("writeFileSync", (handler) => {
+            let args = handler.waitForArguments(...handler.getArgs(0,3));
+            fs.writeFileSync(args.shift(), args.shift(), args.shift());
+            return "";
+        });
+        this.set("mkdirSync", (handler) => {
+            let args = handler.waitForArguments(...handler.getArgs(0,2));
+            fs.mkdirSync(args.shift(), args.shift());
+            return "";
+        });
+        this.set("unlinkSync", (handler) => {
+            let args = handler.waitForArguments(...handler.getArgs(0,1));
+            fs.unlinkSync(args.shift());
+            return "";
+        });
+        this.set("existsSync", (handler) => {
+            let args = handler.waitForArguments(...handler.getArgs(0,1));
+            return fs.existsSync(args.shift());
+        });
+        this.set("statSync", (handler) => {
+            let args = handler.waitForArguments(...handler.getArgs(0,2));
+            return fs.statSync(args.shift(), args.shift());
+        });
 
         this.cache.lock();
     }
@@ -89,15 +166,15 @@ class Process extends Environments.ProtectedEnvironment {
         super();
         
         this.set("print", (handler) => {
-            console.log(...handler.waitForArguments(...handler.getArgs(0,handler.getArgLength())));
+            console.log(...handler.waitForArguments(...handler.getArgs(0,handler.argLength)));
             return "";
         });
         this.set("warn", (handler) => {
-            console.warn(...handler.waitForArguments(...handler.getArgs(0,handler.getArgLength())));
+            console.warn(...handler.waitForArguments(...handler.getArgs(0,handler.argLength)));
             return "";
         });
         this.set("error", (handler) => {
-            console.error(...handler.waitForArguments(...handler.getArgs(0,handler.getArgLength())));
+            console.error(...handler.waitForArguments(...handler.getArgs(0,handler.argLength)));
             return "";
         });
         this.set("memoryUsage", () => process.memoryUsage());
@@ -108,6 +185,10 @@ class Process extends Environments.ProtectedEnvironment {
         this.set("uptime", () => process.uptime());
         this.set("version", process.version);
         this.set("cwd", process.cwd());
+        this.set("write", (handler) => {
+            process.stdin.write(handler.waitForArguments(handler.getArg(0)).shift());
+            return "";
+        });
 
         this.cache.lock();
     }
@@ -118,29 +199,30 @@ class ObjectInteract extends Environments.ProtectedEnvironment {
         super();
 
         this.set("new", (handler) => {
-            if (handler.getArgLength() > 0) {
-                let args = handler.waitForArguments(...handler.getArgs(0, handler.getArgLength()));
+            if (handler.argLength > 0) {
+                let args = handler.waitForArguments(...handler.getArgs(0, handler.argLength));
                 if (args[0]?.constructor) return new (args.shift() as any)(...args)
             } else return {};
         });
         
         this.set("get", (handler) => {
-            if (handler.getArgLength() < 2) throw new Error("Expected 2 argument, got 0");
+            if (handler.argLength < 2) throw new Error("Expected 2 argument, got 0");
             let args = handler.waitForArguments(...handler.getArgs(0, 2));
             return args.shift()[args.shift()];
         });
 
         this.set("set", (handler) => {
-            if (handler.getArgLength() < 3) throw new Error("Expected 3 argument, got 0");
-            let args = handler.waitForArguments(...handler.getArgs(0, 3));
-            return args.shift()[args.shift()] = args.shift();
+            if (handler.argLength < 3) throw new Error("Expected 3 argument, got 0");
+            let [obj, key, value] = handler.waitForArguments(...handler.getArgs(0, 3));
+            obj[key] = value;
+            return obj;
         });
 
         this.set("delete", (handler) => {
-            if (handler.getArgLength() < 2) throw new Error("Expected 2 argument, got 0");
-            let args = handler.waitForArguments(...handler.getArgs(0, 2));
-            delete args.shift()[args.shift()];
-            return "";
+            if (handler.argLength < 2) throw new Error("Expected 2 argument, got 0");
+            let [obj, key] = handler.waitForArguments(...handler.getArgs(0, 2));
+            delete obj[key]
+            return obj;
         });
         this.cache.lock();
     }
@@ -165,7 +247,7 @@ class Arithmetics extends Environments.ProtectedEnvironment {
     public constructor() {
         super()
         this.set("sum", (handler) => {
-            let args = handler.waitForArguments(...handler.getArgs(0, handler.getArgLength()));
+            let args = handler.waitForArguments(...handler.getArgs(0, handler.argLength));
             let v = 0;
             while (args.length > 0) {
                 v += Number(args.shift());
@@ -174,7 +256,7 @@ class Arithmetics extends Environments.ProtectedEnvironment {
             return v;
         });
         this.set("sub", (handler) => {
-            let args = handler.waitForArguments(...handler.getArgs(0, handler.getArgLength()));
+            let args = handler.waitForArguments(...handler.getArgs(0, handler.argLength));
             let v = 0;
             while (args.length > 0) {
                 v -= Number(args.shift());
@@ -183,7 +265,7 @@ class Arithmetics extends Environments.ProtectedEnvironment {
             return v;
         });
         this.set("multi", (handler) => {
-            let args = handler.waitForArguments(...handler.getArgs(0, handler.getArgLength()));
+            let args = handler.waitForArguments(...handler.getArgs(0, handler.argLength));
             let v = 1;
             while (args.length > 0) {
                 v *= Number(args.shift());
@@ -192,7 +274,7 @@ class Arithmetics extends Environments.ProtectedEnvironment {
             return v;
         });
         this.set("div", (handler) => {
-            let args = handler.waitForArguments(...handler.getArgs(0, handler.getArgLength()));
+            let args = handler.waitForArguments(...handler.getArgs(0, handler.argLength));
             let v = Number(args.shift());
             while (args.length > 0) {
                 v /= Number(args.shift());
@@ -233,11 +315,11 @@ class Arithmetics extends Environments.ProtectedEnvironment {
             return Math.trunc(args.shift());
         });
         this.set("min", (handler) => {
-            let args = handler.waitForArguments(...handler.getArgs(0, handler.getArgLength()));
+            let args = handler.waitForArguments(...handler.getArgs(0, handler.argLength));
             return Math.min(...args);
         });
         this.set("max", (handler) => {
-            let args = handler.waitForArguments(...handler.getArgs(0, handler.getArgLength()));
+            let args = handler.waitForArguments(...handler.getArgs(0, handler.argLength));
             return Math.max(...args);
         });
         this.set("pow", (handler) => {
@@ -257,7 +339,9 @@ const Modules = {
     Arithmetics,
     Process,
     ObjectInteract,
-    OS
+    OS,
+    FileSystem,
+    Path
 }
 
 export {