+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## v1.0.10
+### Added / New
+- Added `Path`, `FileSystem` class
+- New features `$eval`
+- getter method `argLength`, a replacement for `getArgLength`
+### Deprecated
+- Function `getArgLength` deprecated, use `argLength` instead
 ## v1.0.9
 - Added `OS`, `Process`, `ObjectInteract` class
 - New features `$throw`, `$new`, `$tryAndCatch`, `$let`, `$get`